javaagent: Handle containers that use custom MBeanServers
Some containers (e.g. Wildfly 8, 9, 10) use a custom MBeanServer.
The simplest way to deal with this is to ask the user to set the factory
to either the correct class (if known) or a non-existant class and let the
container intialisation process correct it.

Any call to Management.getPlatformMBeanServer() before the container is
ready will then fail with a ClassNotFoundException until container
initialisation is complete.

In terms of scraping this will cause an error until the container is ready (probably an error 500 or similar)

Fixes #128