Update Cassandra example config
Cassandra has added a lot more metrics recently. The existing example
config results in a scrape time of 15-30 seconds and more than doubles
Cassandra's CPU usage in our setup.

This changes the example config to use a whitelist approach with a
sample whitelist which takes a much more reasonable 200 ms to scrape.

Signed-off-by: Russ Garrett <russ@garrett.co.uk>